reorganized stmt in similar way to expr
