+++ src/ast/stmt/stmt_kind.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, Copy)]
+pub enum StmtKind {
+    Var,
+    Function,
+}
